pyproject.toml
[project]
name = "shared-garden-model"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastai==2.8.6",
    "gradio>=6.2.0",
    "ipython>=8.37.0",
    "pillow>=12.0.0",
    "torch>=2.9.1",
    "torchvision>=0.24.1",
]
